requirements.txt
python-dotenv
utils
python-magic

streamlit


google-api-python-client
requests



pydantic
langchain

mistralai

pymupdf
peft

llama-index

llama-index-readers-file
llama-index-core
llama-parse

llama-index-llms-mistralai
llama-index-llms-langchain


llama-index-postprocessor-cohere-rerank
llama-index-llms-cohere
llama-index-embeddings-cohere

llama-index-embeddings-huggingface
llama-index-embeddings-mistralai

llama-index-embeddings-openai
llama-index-postprocessor-flag-embedding-reranker

llama-index-postprocessor-cohere-rerank
llama-index-embeddings-cohere

crewai
openai

graspologic
numpy
scipy

reportlab

weasyprint
matplotlib
jinja2